requirements.txt
alembic==1.18.3
blinker==1.9.0
cachelib==0.13.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==25.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
msgspec==0.20.0
packaging==26.0
pyaes==1.6.1
pyasn1==0.6.2
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
SQLAlchemy==2.0.46
Telethon==1.42.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
APScheduler==3.10.4
trino[external-authentication]==0.329.0
pandas
openpyxl
